+++ .env.example
@@ -1,0 +1,2 @@
+# Set the backend API key for OpenAI, so that users don't have to provide one
+OPENAI_API_KEY=
