+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nextjs-chatgpt-app",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -24,7 +24,7 @@
     "zustand": "^4.3.6"
   },
   "devDependencies": {
-    "@types/node": "^18.15.8",
+    "@types/node": "^18.15.10",
     "@types/prismjs": "^1.26.0",
     "@types/react": "^18.0.29",
     "@types/react-dom": "^18.0.11",
