+++ pages/_app.tsx
@@ -1,60 +1,15 @@
 import * as React from 'react';
 import Head from 'next/head';
-import createCache from '@emotion/cache';
 import { Analytics as VercelAnalytics } from '@vercel/analytics/react';
 import { AppProps } from 'next/app';
 import { CacheProvider, EmotionCache } from '@emotion/react';
-import { CssBaseline, CssVarsProvider, extendTheme } from '@mui/joy';
-import { Inter, JetBrains_Mono } from 'next/font/google';
+import { CssBaseline, CssVarsProvider } from '@mui/joy';
 
-
-// Application Theme (JoyUI = MUI) & Fonts (Next.js)
-
-const inter = Inter({
-  weight: ['400', '500', '600', '700'],
-  subsets: ['latin'],
-  display: 'swap',
-  fallback: ['Helvetica', 'Arial', 'sans-serif'],
-});
-
-const jetBrainsMono = JetBrains_Mono({
-  weight: ['400', '500', '600', '700'],
-  subsets: ['latin'],
-  display: 'swap',
-  fallback: ['monospace'],
-});
-
-const theme = extendTheme({
-  fontFamily: {
-    body: inter.style.fontFamily,
-    code: jetBrainsMono.style.fontFamily,
-  },
-});
-
-export const bodyFontClassName = inter.className;
+import { createEmotionCache, theme } from '@/lib/theme';
 
 
 // Client-side cache, shared for the whole session of the user in the browser.
-
-const isBrowser = typeof document !== 'undefined';
-
-export function createEmotionCache() {
-  let insertionPoint;
-
-  if (isBrowser) {
-    // On the client side, _document.tsx has a meta tag with the name "emotion-insertion-point" at the top of the <head>.
-    // This assures that MUI styles are loaded first, and allows allows developers to easily override MUI styles with other solutions like CSS modules.
-    const emotionInsertionPoint = document.querySelector<HTMLMetaElement>(
-      'meta[name="emotion-insertion-point"]',
-    );
-    insertionPoint = emotionInsertionPoint ?? undefined;
-  }
-
-  return createCache({ key: 'mui-style', insertionPoint });
-}
-
 const clientSideEmotionCache = createEmotionCache();
-
 
 export interface MyAppProps extends AppProps {
   emotionCache?: EmotionCache;
